move tabbed shelf to new architecture
